+++ examples/typescript-react-example/src/InnerModuleSignature.gen.tsx
@@ -1,0 +1,13 @@
+/* TypeScript file generated from InnerModuleSignature.re by genType. */
+/* eslint-disable import/first */
+
+
+// tslint:disable-next-line:no-var-requires
+const InnerModuleSignatureBS = require('./InnerModuleSignature.bs');
+
+// tslint:disable-next-line:interface-over-type-literal
+export type InnerModule_t = string;
+
+export const InnerModule_make: (_1:InnerModule_t) => string = InnerModuleSignatureBS.InnerModule.InnerModule;
+
+export const InnerModule: { make: (_1:InnerModule_t) => string } = InnerModuleSignatureBS.InnerModule