+++ Changes.md
@@ -1,4 +1,5 @@
 # master
+- Add missing support for annotations in inner module signatures (e.g. `module M : { ... } = { ... }`).
 
 # 3.37.0
 - Fix issue where a conversion was not applied to optional arguments due to mistaken covariance.